Move subgraph url to own module
