+++ src/hooks/useTokenRecords.ts
@@ -1,6 +1,6 @@
-import { getSubgraphUrl } from "src/constants";
 import { useTokenRecordsQuery } from "src/generated/graphql";
 import { getTreasuryAssetValue } from "src/helpers/subgraph/TreasuryQueryHelper";
+import { getSubgraphUrl } from "src/helpers/SubgraphUrlHelper";
 import { DEFAULT_RECORD_COUNT } from "src/views/TreasuryDashboard/components/Graph/Constants";
 
 const QUERY_OPTIONS = { refetchInterval: 60000 }; // Refresh every 60 seconds
